Chore: Move project writeup content to README.md
The README.md file is among the first files to be seen when someone visits
a repository on Github. We had a project writeup, but it was not automatically
displayed on the repo's front page. This commit seeks to move the writeup
contents, which includes setup instructions and project context, to the new
README.md file.

Addresses project visibility and transparency.